feat: Add explanation for using `append` method in Rails
Add a new blog post explaining how to use the `append` method in Ruby on Rails to add `before_save` callbacks to classes in external gems. The post provides examples of reopening classes and using `class_eval` or `prepend` to achieve this functionality. It also highlights important points to consider, such as potential issues with gem updates and performance impact.

Closes #123